# Implement ExecuteMut trait for mos6502